Refactored so that GameManager completely controls what happens when the game starts, and the click event callback in MainMenuManager just calls this

+++ Assets/Scripts/UI/GameUIManager.cs
@@ -9,7 +9,7 @@ public class GameUIManager : UIManger
     private Label _timer;
     private Label _heartCounter;
     private float _timeLeft;
-    private float _heartCount;
+    private int _heartCount;
     public static GameUIManager Instance { get; private set; }
     
     protected override void Awake()
